# backend/app/api/afnor/v1.py
"""API AFNOR XP Z12-013 exposée à Odoo — version v1 (spec.md § 4.4/§ 4.8).

Seule version existante à ce jour (la norme XP Z12-013 n'a pas encore de v2
publiée) — point d'entrée du registre de versions
(`app/afnor/versioning/registry.py`), qui permettra d'en monter une future sur un
autre préfixe sans toucher à celui-ci ni aux services qu'il appelle. Les routes
métier communes à toute future version (`POST /flows`, `POST /flows/search`,
`GET /flows/{flowId}`, `GET /siren/code-insee:{siren}`) sont factorisées dans
`_common.py` ; seule `/oauth/token` (infrastructure non versionnée, § 4.10) est
propre à ce module."""

from fastapi import APIRouter, Depends, Form, Request, Response
from sqlalchemy.orm import Session

from app.afnor.versioning.registry import register_version
from app.api.afnor._common import register_common_routes
from app.auth.oauth import issue_token_response
from app.auth.rate_limit import rate_limit
from app.db.session import get_db
from app.services import audit_trace_service

AFNOR_API_VERSION = "v1"

router = APIRouter()
register_common_routes(router, AFNOR_API_VERSION)


@router.post("/oauth/token", dependencies=[Depends(rate_limit(scope="oauth_token"))])
def issue_token(
    request: Request,
    response: Response,
    grant_type: str = Form(...),
    client_id: str = Form(...),
    client_secret: str = Form(...),
    db: Session = Depends(get_db),
):
    """Conformité RFC 6749 (authentification client, validation du grant, erreurs
    normalisées) déléguée à Authlib — cf. `app.auth.oauth.issue_token_response`."""
    status_code, body = issue_token_response(
        db, grant_type=grant_type, client_id=client_id, client_secret=client_secret
    )
    response.status_code = status_code
    audit_trace_service.record_odoo_flow_trace(
        db,
        request,
        afnor_api_version=AFNOR_API_VERSION,
        endpoint="POST /oauth/token",
        client_id=client_id,
        response={"status": "issued" if status_code == 200 else body.get("error")},
        http_status=status_code,
    )
    return body


register_version(AFNOR_API_VERSION, router)